Move private classes to internal package
Finer-grained  change in moving private API from public to internal
package, this time for various packages. Other than moving the classes,
fixing checkstyle, adding visibility modifiers and formatting, there is
no code or behaviour change.

Classes moved:
com.hazelcast.cache.journal.EventJournalCacheEvent

com.hazelcast.instance.HazelcastInstanceCacheManager

com.hazelcast.cluster.memberselector.AndMemberSelector
com.hazelcast.cluster.memberselector.OrMemberSelector

com.hazelcast.config.DeclarativeConfigUtil

com.hazelcast.logging.LoggingServiceImpl

com.hazelcast.query.Metadata

com.hazelcast.cache.hidensity.impl.nativememory.CacheHiDensityRecordProcessor
com.hazelcast.cache.hidensity.impl.nativememory.HiDensityNativeMemoryCacheEntryProcessorEntry
com.hazelcast.cache.hidensity.impl.nativememory.HiDensityNativeMemoryCacheRecord
com.hazelcast.cache.hidensity.impl.nativememory.HiDensityNativeMemoryCacheRecordAccessor
com.hazelcast.cache.hidensity.impl.nativememory.HiDensityNativeMemoryCacheRecordMap
com.hazelcast.cache.hidensity.impl.nativememory.HiDensityNativeMemoryCacheRecordStore
com.hazelcast.cache.hidensity.impl.nativememory.HotRestartHiDensityNativeMemoryCacheRecordMap
com.hazelcast.cache.hidensity.impl.nativememory.HotRestartHiDensityNativeMemoryCacheRecordStore

com.hazelcast.cache.hidensity.impl.HiDensityCacheRecord

com.hazelcast.cache.hidensity.maxsize.HiDensityEntryCountEvictionChecker
com.hazelcast.cache.hidensity.maxsize.HiDensityFreeNativeMemoryPercentageEvictionChecker
com.hazelcast.cache.hidensity.maxsize.HiDensityFreeNativeMemorySizeEvictionChecker
com.hazelcast.cache.hidensity.maxsize.HiDensityUsedNativeMemoryPercentageEvictionChecker
com.hazelcast.cache.hidensity.maxsize.HiDensityUsedNativeMemorySizeEvictionChecker

com.hazelcast.cache.hidensity.operation.BackupAwareHiDensityCacheOperation
com.hazelcast.cache.hidensity.operation.BackupAwareKeyBasedHiDensityCacheOperation
com.hazelcast.cache.hidensity.operation.CacheBackupRecordStore
com.hazelcast.cache.hidensity.operation.CacheContainsKeyOperation
com.hazelcast.cache.hidensity.operation.CacheEntryProcessorOperation
com.hazelcast.cache.hidensity.operation.CacheGetAllOperation
com.hazelcast.cache.hidensity.operation.CacheGetAllOperationFactory
com.hazelcast.cache.hidensity.operation.CacheGetAndRemoveOperation
com.hazelcast.cache.hidensity.operation.CacheGetAndReplaceOperation
com.hazelcast.cache.hidensity.operation.CacheGetOperation
com.hazelcast.cache.hidensity.operation.CacheLoadAllOperation
com.hazelcast.cache.hidensity.operation.CacheLoadAllOperationFactory
com.hazelcast.cache.hidensity.operation.CacheMergeBackupOperation
com.hazelcast.cache.hidensity.operation.CacheMergeOperation
com.hazelcast.cache.hidensity.operation.CacheMergeOperationFactory
com.hazelcast.cache.hidensity.operation.CachePutAllBackupOperation
com.hazelcast.cache.hidensity.operation.CachePutAllOperation
com.hazelcast.cache.hidensity.operation.CachePutBackupOperation
com.hazelcast.cache.hidensity.operation.CachePutIfAbsentOperation
com.hazelcast.cache.hidensity.operation.CachePutOperation
com.hazelcast.cache.hidensity.operation.CacheRemoveBackupOperation
com.hazelcast.cache.hidensity.operation.CacheRemoveOperation
com.hazelcast.cache.hidensity.operation.CacheReplaceOperation
com.hazelcast.cache.hidensity.operation.CacheSegmentShutdownOperation
com.hazelcast.cache.hidensity.operation.CacheSetExpiryPolicyBackupOperation
com.hazelcast.cache.hidensity.operation.CacheSetExpiryPolicyOperation
com.hazelcast.cache.hidensity.operation.CacheSizeOperation
com.hazelcast.cache.hidensity.operation.CacheSizeOperationFactory
com.hazelcast.cache.hidensity.operation.HiDensityCacheDataSerializerHook
com.hazelcast.cache.hidensity.operation.HiDensityCacheOperation
com.hazelcast.cache.hidensity.operation.HiDensityCacheOperationProvider
com.hazelcast.cache.hidensity.operation.HiDensityCacheReplicationOperation
com.hazelcast.cache.hidensity.operation.KeyBasedHiDensityCacheOperation
com.hazelcast.cache.hidensity.operation.WanCacheMergeOperation
com.hazelcast.cache.hidensity.operation.WanCacheRemoveOperation

com.hazelcast.cache.hidensity.HiDensityCacheRecordStore
com.hazelcast.cache.hidensity.HiDensityCacheStorageInfo
com.hazelcast.cache.hidensity.SampleableHiDensityCacheRecordMap

com.hazelcast.cache.hotrestart.HotRestartEnterpriseCacheRecordStore

com.hazelcast.enterprise.monitor.impl.jmx.EnterpriseManagementService
com.hazelcast.enterprise.monitor.impl.jmx.LicenseInfoMBean

com.hazelcast.enterprise.monitor.impl.management.EnterpriseManagementCenterConnectionFactory
com.hazelcast.enterprise.monitor.impl.management.EnterpriseNodeStateImpl
com.hazelcast.enterprise.monitor.impl.management.EnterpriseTimedMemberStateFactory

com.hazelcast.enterprise.monitor.impl.rest.EnterpriseHttpGetCommandProcessor
com.hazelcast.enterprise.monitor.impl.rest.EnterpriseTextCommandServiceImpl
com.hazelcast.enterprise.monitor.impl.rest.LicenseInfoImpl

com.hazelcast.enterprise.monitor.LicenseInfo

com.hazelcast.enterprise.EnterprisePhoneHome

com.hazelcast.nio.serialization.EnterpriseSerializationService

com.hazelcast.nio.ssl.AbstractMultiSocketTLSChannelInitializer
com.hazelcast.nio.ssl.AbstractTLSChannelInitializer
com.hazelcast.nio.ssl.ChannelHandlerPair
com.hazelcast.nio.ssl.ClientTLSChannelInitializer
com.hazelcast.nio.ssl.MemberTLSChannelInitializer
com.hazelcast.nio.ssl.SSLEngineFactoryAdaptor
com.hazelcast.nio.ssl.SSLEngineFactorySupport
com.hazelcast.nio.ssl.TextTLSChannelInitializer
com.hazelcast.nio.ssl.TLSDecoder
com.hazelcast.nio.ssl.TLSEncoder
com.hazelcast.nio.ssl.TLSExecutor
com.hazelcast.nio.ssl.TLSHandshakeDecoder
com.hazelcast.nio.ssl.TLSHandshakeEncoder
com.hazelcast.nio.ssl.TLSUtil
com.hazelcast.nio.ssl.UnifiedTLSChannelInitializer

com.hazelcast.nio.CipherByteArrayProcessor
com.hazelcast.nio.CipherHelper
com.hazelcast.nio.EnterpriseBufferObjectDataInput
com.hazelcast.nio.EnterpriseBufferObjectDataOutput
com.hazelcast.nio.EnterpriseChannelInitializerProvider
com.hazelcast.nio.EnterpriseObjectDataInput
com.hazelcast.nio.EnterpriseObjectDataOutput

com.hazelcast.security.ClusterCallbackHandler
com.hazelcast.security.EmptyParametersImpl
com.hazelcast.security.LoginConfigurationDelegate
com.hazelcast.security.ParametersImpl
com.hazelcast.security.SecureCallableImpl
com.hazelcast.security.SecurityConstants
com.hazelcast.security.SecurityContextImpl
com.hazelcast.security.SecurityUtil
