+++ hazelcast/src/main/java/com/hazelcast/cache/impl/HazelcastInstanceCacheManager.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.instance;
+package com.hazelcast.cache.impl;
 
 import com.hazelcast.cache.HazelcastCacheManager;
 import com.hazelcast.cache.ICache;
