+++ hazelcast/src/main/java/com/hazelcast/cluster/memberselector/impl/AndMemberSelector.java
@@ -14,30 +14,31 @@
  * limitations under the License.
  */
 
-package com.hazelcast.cluster.memberselector;
+package com.hazelcast.cluster.memberselector.impl;
 
 import com.hazelcast.cluster.Member;
 import com.hazelcast.cluster.MemberSelector;
 
 /**
- * Selects a member if one of the sub-selectors succeed
+ * Selects a member only if all of the sub-selectors succeed
  */
-class OrMemberSelector implements MemberSelector {
+public class AndMemberSelector
+        implements MemberSelector {
 
     private final MemberSelector[] selectors;
 
-    OrMemberSelector(MemberSelector... selectors) {
+    public AndMemberSelector(MemberSelector... selectors) {
         this.selectors = selectors;
     }
 
     @Override
     public boolean select(Member member) {
         for (MemberSelector selector : selectors) {
-            if (selector.select(member)) {
-                return true;
+            if (!selector.select(member)) {
+                return false;
             }
         }
 
-        return false;
+        return true;
     }
 }
