+++ hazelcast/src/main/java/com/hazelcast/config/AbstractConfigLocator.java
@@ -28,8 +28,8 @@ import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.Collection;
 
-import static com.hazelcast.config.DeclarativeConfigUtil.isAcceptedSuffixConfigured;
-import static com.hazelcast.config.DeclarativeConfigUtil.throwUnacceptedSuffixInSystemProperty;
+import static com.hazelcast.internal.config.DeclarativeConfigUtil.isAcceptedSuffixConfigured;
+import static com.hazelcast.internal.config.DeclarativeConfigUtil.throwUnacceptedSuffixInSystemProperty;
 import static com.hazelcast.internal.util.Preconditions.checkFalse;
 import static java.util.Objects.requireNonNull;
