+++ hazelcast/src/main/java/com/hazelcast/config/XmlConfigLocator.java
@@ -16,8 +16,8 @@
 
 package com.hazelcast.config;
 
-import static com.hazelcast.config.DeclarativeConfigUtil.SYSPROP_MEMBER_CONFIG;
-import static com.hazelcast.config.DeclarativeConfigUtil.XML_ACCEPTED_SUFFIXES;
+import static com.hazelcast.internal.config.DeclarativeConfigUtil.SYSPROP_MEMBER_CONFIG;
+import static com.hazelcast.internal.config.DeclarativeConfigUtil.XML_ACCEPTED_SUFFIXES;
 
 /**
  * Support class for the {@link XmlConfigBuilder} that locates the XML configuration:
