+++ hazelcast/src/test/java/com/hazelcast/client/config/YamlClientConfigBuilderResolutionTest.java
@@ -30,8 +30,8 @@ import org.junit.runner.RunWith;
 
 import java.io.File;
 
-import static com.hazelcast.config.DeclarativeConfigUtil.SYSPROP_CLIENT_CONFIG;
-import static com.hazelcast.config.DeclarativeConfigUtil.YAML_ACCEPTED_SUFFIXES_STRING;
+import static com.hazelcast.internal.config.DeclarativeConfigUtil.SYSPROP_CLIENT_CONFIG;
+import static com.hazelcast.internal.config.DeclarativeConfigUtil.YAML_ACCEPTED_SUFFIXES_STRING;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(HazelcastSerialClassRunner.class)
